Move main package to src folder
