+++ src/detritalPy/GSA_2020_short_course/ExampleDataset_1.xlsx
