+++ src/detritalPy/GSA_2020_short_course/INPUT_Blum et al. (2018) Scientific Reports.xlsx
