+++ src/detritalPy/GSA_2020_short_course/INPUT_Gehrels et al. (2020) GChron.xlsx
