+++ src/detritalPy/GSA_2022_short_course/ExampleDataset_1.xlsx
