+++ src/detritalPy/GSA_2022_short_course/INPUT_Gehrels et al. (2020) GChron.xlsx
