+++ src/detritalPy/adaptiveKDE.py
