+++ src/detritalPy/example-data/Blum et al. (2018) Scientific Reports.xlsx
