+++ src/detritalPy/example-data/Sundell and Saylor (2017) ds02.xlsx
