+++ src/detritalPy/nmfFuncs.py
